+++ R/arcgisutils-package.R
@@ -1,5 +1,6 @@
 ## usethis namespace: start
 #' @importFrom lifecycle deprecated
+#' @importFrom s7x Enum
 #' @keywords internal
 #' @useDynLib arcgisutils, .registration = TRUE
 ## usethis namespace: end